Refactor playground
