--- universal-login-react/src/config/types.ts
@@ -1,10 +1,0 @@
-import {IPGeolocationApiConfig} from '@unilogin/commons';
-
-export interface Config {
-  domains: string[];
-  relayerUrl: string;
-  jsonRpcUrl: string;
-  tokens: string[];
-  saiTokenAddress?: string;
-  ipGeolocationApi: IPGeolocationApiConfig;
-}
