--- universal-login-react/src/core/services/useServices.ts
@@ -1,6 +1,0 @@
-import {useContext} from 'react';
-import {ServiceContext} from './createServices';
-
-export const useServices = () => {
-  return useContext(ServiceContext);
-};
