+++ universal-login-react/src/playground/config.ts
@@ -1,4 +1,13 @@
-import {Config} from './types';
+import {IPGeolocationApiConfig} from '@unilogin/commons';
+
+export interface Config {
+  domains: string[];
+  relayerUrl: string;
+  jsonRpcUrl: string;
+  tokens: string[];
+  saiTokenAddress?: string;
+  ipGeolocationApi: IPGeolocationApiConfig;
+}
 
 require('dotenv').config();
 import {ETHER_NATIVE_TOKEN} from '@unilogin/commons';
