+++ universal-login-react/src/playground/main.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import {render} from 'react-dom';
+import App from './App';
+
+render(<App/>, document.getElementById('app'));
